consoletwitter/src/test/java/ro/ghionoiu/twitter/command/handlers/FollowCommandHandlerTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package ro.ghionoiu.twitter.command.handlers;

import org.junit.Test;
import static org.junit.Assert.*;
import static org.hamcrest.CoreMatchers.*;
import org.junit.Before;
import ro.ghionoiu.twitter.command.CommandHandler;

/**
 *
 * @author Iulian Ghionoiu <iulian.ghionoiu@exenne.ro>
 */
public class FollowCommandHandlerTest {
    private CommandHandler instance;
    
    @Before
    public void setUp() {
        instance = new FollowCommandHandler();
    }
    
    @Test
    public void can_handle_commands_that_contain_follows() {
        
        boolean canHandle = instance.canHandle("Alice follows Bob");
        
        assertThat(canHandle, is(true));
    }
    
    @Test
    public void cannot_handle_commands_that_do_not_contain_follows() {
        
        boolean canHandle = instance.canHandle("Alice");
        
        assertThat(canHandle, is(false));
    }
}